requirements.txt
python-dotenv==1.0.0
requests==2.28.1
tqdm==4.64.1